64-bit to 32-bit conversion warnings removed

+++ Classes/AQGridViewUpdateItem.m
@@ -128,7 +128,7 @@
 - (NSUInteger) index
 {
 	// handle case where offset is negative and would cause index to wrap
-	if ( (_offset < 0) && (abs(_offset) > _index) )
+	if ( (_offset < 0) && (abs((int)_offset) > _index) )
 		return ( 0 );
 	
 	return ( _index + _offset );
